+++ src/ert/config/field_config.py
@@ -28,7 +28,7 @@ _logger = logging.getLogger(__name__)
 
 
 @dataclass
-class Field(ParameterConfig):
+class FieldConfig(ParameterConfig):
     nx: int
     ny: int
     nz: int
