+++ tests/unit_tests/config/config_dict_generator.py
@@ -23,7 +23,7 @@ from ert.config import (
     queue_string_options,
 )
 from ert.config._read_summary import make_summary_key
-from ert.config.field import TRANSFORM_FUNCTIONS
+from ert.config.field_config import TRANSFORM_FUNCTIONS
 from ert.config.parsing import ConfigKeys, HistorySource
 
 from .egrid_generator import EGrid, egrids
